loader: comparison between signed and unsigned integer expressions
/code/illumos-gate/usr/src/boot/lib/libstand/net.c:116:23: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
    if (cc != -1 && cc < ssize)
                       ^
cc1: all warnings being treated as errors

+++ usr/src/boot/lib/libstand/net.c
@@ -113,7 +113,7 @@ sendrecv(struct iodesc *d,
 				return (-1);
 			}
 			cc = (*sproc)(d, sbuf, ssize);
-			if (cc != -1 && cc < ssize)
+			if (cc != -1 && cc < (ssize_t)ssize)
 				panic("sendrecv: short write! (%zd < %zd)",
 				    cc, ssize);
